worktree: refactor infer_backlink() to use strbuf*
The infer_backlink() function initializes a `strbuf` for the inferred
backlink and returns the result as a `char*` by detaching the `strbuf`.

The next patch needs the backlink returned from infer_backlink() as a
`strbuf`. It seemed inefficient to convert from `strbuf` to `char*`
and back to `strbuf` again. This refactors infer_backlink() to return
an integer result and accept a pre-allocated `strbuf` for the inferred
backlink path, improving efficiency.

Signed-off-by: Caleb White <cdwhite3@pm.me>