worktree: add id to `worktree list` output
The worktree id is relatively hidden from the user, however, there may
be times where a user or script needs to determine the worktree id for a
linked worktree (e.g., to manually operate on some refs). While the id
is stored in the worktree `.git` file, it would be nice if there was an
easier method of obtaining it. This teaches Git to output the worktree id
(for linked worktrees) in the `worktree list` verbose and porcelain modes.

Signed-off-by: Caleb White <cdwhite3@pm.me>